fix(interceptor): mypy strict — type wrappers, latency_ms float, drop dead type-ignores
- RawEvent.latency_ms: int → float (latency is naturally fractional)
- MetricEntry.total/p50/p95_latency_ms: int → float to match
- _Bucket.latencies: List[int] → List[float]; _compute_percentile updated
- _patched_urlopen: explicit Any-annotated signature, _UNSET sentinel for
  lazy timeout resolution, remove dead pool_connections/pool_maxsize kwargs
- _patched_send, _patched_async_send, _patched_request: explicit Any signatures
- Replace # type: ignore[no-untyped-def] with proper signatures throughout
- Replace # type: ignore[attr-defined] with # type: ignore[method-assign] on
  all method-assignment patching lines (urllib3, httpx, aiohttp)
- Remove now-stale # type: ignore[import-untyped] comments (urllib3/aiohttp
  ship with inline types when installed)

+++ recost/_types.py
@@ -42,7 +42,7 @@ class RawEvent:
     host: str
     path: str
     status_code: int
-    latency_ms: int
+    latency_ms: float
     request_bytes: int
     response_bytes: int
     provider: Optional[str] = None
@@ -63,9 +63,9 @@ class MetricEntry:
     method: str
     request_count: int
     error_count: int
-    total_latency_ms: int
-    p50_latency_ms: int
-    p95_latency_ms: int
+    total_latency_ms: float
+    p50_latency_ms: float
+    p95_latency_ms: float
     total_request_bytes: int
     total_response_bytes: int
     estimated_cost_cents: float
